# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: ShelterFlex Backend API
  description: API for ShelterFlex backend service
  version: 0.1.0
  contact:
    name: ShelterFlex Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.shelterflex.io
    description: Production server
paths:
...
  /api/auth/request-otp:
    post:
      summary: Request OTP for email login
      description: Sends a one-time password (OTP) to the provided email address.
      operationId: requestOtp
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: OTP request accepted
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    example: OTP sent to your email
        "429":
          $ref: "#/components/responses/TooManyRequests"
        default:
          $ref: "#/components/responses/Error"
  /api/auth/verify-otp:
    post:
      summary: Verify OTP and create session
      description: Verifies the OTP for the email and returns a session token (Bearer).
      operationId: verifyOtp
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - otp
              properties:
                email:
                  type: string
                  format: email
                otp:
                  type: string
                  minLength: 6
                  maxLength: 6
      responses:
        "200":
          description: OTP verified successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - token
                  - user
                properties:
                  token:
                    type: string
                    description: "Session token to be used as Authorization: Bearer <token>"
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
  /api/auth/logout:
    post:
      summary: Logout
      description: Invalidates the current session token.
      operationId: logout
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    example: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Error"
  /api/auth/me:
    get:
      summary: Get current authenticated user
      description: Returns the current user associated with the Bearer session token.
      operationId: getMe
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Error"
...
  /api/auth/wallet/challenge:
    post:
      summary: Create wallet login challenge
      description: Returns a Stellar SEP-10 style challenge XDR for the wallet to sign.
      operationId: walletChallenge
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletChallengeRequest'
      responses:
        "200":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletChallengeResponse'
        "400":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
  /api/auth/wallet/verify:
    post:
      summary: Verify wallet login signature
      description: Verifies a signed challenge XDR and creates a session token.
      operationId: walletVerify
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletVerifyRequest'
      responses:
        "200":
          description: Wallet verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        "400":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
...
components:
...
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - createdAt
        - name
        - role
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        name:
          type: string
        role:
          type: string
          enum: [tenant, landlord, agent]
    AuthTokenResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    WalletChallengeRequest:
      type: object
      required:
        - address
      properties:
        address:
          type: string
          description: Stellar public key (G...)
          pattern: '^G[A-Z0-9]{55}$'
    WalletChallengeResponse:
      type: object
      required:
        - challengeXdr
        - expiresAt
      properties:
        challengeXdr:
          type: string
        expiresAt:
          type: string
          format: date-time
    WalletVerifyRequest:
      type: object
      required:
        - address
        - signedChallengeXdr
      properties:
        address:
          type: string
          description: Stellar public key (G...)
          pattern: '^G[A-Z0-9]{55}$'
        signedChallengeXdr:
          type: string
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - TOO_MANY_REQUESTS
                - NOT_FOUND
                - CONFLICT
                - LISTING_ALREADY_RENTED
                - SOROBAN_ERROR
                - INTERNAL_ERROR
              example: NOT_FOUND
            message:
              type: string
              example: "Resource not found"
            details:
              type: object
              description: Optional structured context for the error
              additionalProperties: true
...
  responses:
    Error:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "TOO_MANY_REQUESTS"
              message: "Rate limit exceeded"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              details:
                fieldErrors:
                  - field: "email"
                    message: "Invalid email address"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: >-
        Ordinary authenticated-session bearer token (authenticateToken middleware only —
        no elevated role/permission required).
    bearerAdminAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: >-
        Same bearer token as bearerAuth, but the route additionally requires the caller's
        session to carry an admin-tier RBAC permission (requirePermission(resource, action)
        in middleware/rbac.ts, e.g. payouts:trigger) or an explicit admin/super_admin role
        check. Kept as a distinct scheme name so admin-gated operations are visibly
        distinguishable from ordinary authenticated ones in this document.
    adminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
      description: >-
        Shared operator secret checked against MANUAL_ADMIN_SECRET. Used by the
        job-scheduler admin routes, which are reachable without a user session.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: >-
        Provider-specific HMAC signature header verified server-side instead of a bearer
        token (e.g. x-paystack-signature for Paystack, verif-hash for Flutterwave). The
        header name varies by payment rail; see the payments webhook handler.
...